+++ generators/app/config/paths.js
@@ -15,5 +15,6 @@ module.exports = {
   },
   pkg: {
     version: '<%= pkg.version %>'
-  }
+  },
+  openveo: 'node_modules/@openveo/'
 };